refactor: 优化了部分代码
