+++ src-tauri/src/core/device/linux.rs
@@ -244,7 +244,7 @@ fn build_device_event(event: &Event) -> Option<DeviceEvent> {
 #[command]
 pub async fn start_device_listening<R: Runtime>(app_handle: AppHandle<R>) -> Result<(), String> {
     if IS_RUNNING.load(Ordering::SeqCst) {
-        return Err("Device is already listening".to_string());
+        return Ok(());
     }
 
     IS_RUNNING.store(true, Ordering::SeqCst);
