+++ src-tauri/src/core/gamepad.rs
@@ -21,7 +21,7 @@ pub struct GamepadEvent {
 #[command]
 pub async fn start_gamepad_listing<R: Runtime>(app_handle: AppHandle<R>) -> Result<(), String> {
     if IS_LISTENING.load(Ordering::SeqCst) {
-        return Err("Gamepad is already listening".to_string());
+        return Ok(());
     }
 
     IS_LISTENING.store(true, Ordering::SeqCst);
