+++ src/pages/main/index.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
-import { convertFileSrc, invoke } from '@tauri-apps/api/core'
+import { convertFileSrc } from '@tauri-apps/api/core'
+import { PhysicalSize } from '@tauri-apps/api/dpi'
 import { Menu } from '@tauri-apps/api/menu'
 import { getCurrentWebviewWindow } from '@tauri-apps/api/webviewWindow'
 import { exists } from '@tauri-apps/plugin-fs'
@@ -10,28 +11,24 @@ import { useDevice } from '@/composables/useDevice'
 import { useGamepad } from '@/composables/useGamepad'
 import { useModel } from '@/composables/useModel'
 import { useSharedMenu } from '@/composables/useSharedMenu'
-import { INVOKE_KEY } from '@/constants'
 import { hideWindow, setAlwaysOnTop, setTaskbarVisibility, showWindow } from '@/plugins/window'
 import { useCatStore } from '@/stores/cat'
 import { useGeneralStore } from '@/stores/general.ts'
 import { useModelStore } from '@/stores/model'
 import { join } from '@/utils/path'
 
+const { startListening } = useDevice()
 const appWindow = getCurrentWebviewWindow()
-const { handleLoad, handleDestroy, handleResize } = useModel()
+const { modelSize, handleLoad, handleDestroy, handleResize, handleKeyChange } = useModel()
 const catStore = useCatStore()
 const { getSharedMenu } = useSharedMenu()
 const modelStore = useModelStore()
 const generalStore = useGeneralStore()
 const resizing = ref(false)
 const backgroundImagePath = ref<string>()
-
-useDevice()
 useGamepad()
 
-onMounted(() => {
-  invoke(INVOKE_KEY.START_DEVICE_LISTENING)
-})
+onMounted(startListening)
 
 onUnmounted(handleDestroy)
 
@@ -59,6 +56,29 @@ watch(() => modelStore.currentModel, async (model) => {
   backgroundImagePath.value = existed ? convertFileSrc(path) : void 0
 }, { deep: true, immediate: true })
 
+watch([() => catStore.scale, modelSize], async () => {
+  if (!modelSize.value) return
+
+  const { width, height } = modelSize.value
+
+  appWindow.setSize(
+    new PhysicalSize({
+      width: Math.round(width * (catStore.scale / 100)),
+      height: Math.round(height * (catStore.scale / 100)),
+    }),
+  )
+}, { immediate: true })
+
+watch(modelStore.pressedKeys, (keys) => {
+  const values = Object.values(keys)
+
+  const hasLeft = values.some(item => item.includes('left-'))
+  const hasRight = values.some(item => item.includes('right-'))
+
+  handleKeyChange(true, hasLeft)
+  handleKeyChange(false, hasRight)
+}, { deep: true })
+
 watch(() => catStore.visible, async (value) => {
   value ? showWindow() : hideWindow()
 })
@@ -68,16 +88,6 @@ watch(() => catStore.penetrable, (value) => {
 }, { immediate: true })
 
 watch(() => catStore.alwaysOnTop, setAlwaysOnTop, { immediate: true })
-
-watch(() => modelStore.currentModel, async (model) => {
-  if (!model) return
-
-  const path = join(model.path, 'resources', 'background.png')
-
-  const existed = await exists(path)
-
-  backgroundImagePath.value = existed ? convertFileSrc(path) : void 0
-}, { deep: true, immediate: true })
 
 watch(() => generalStore.taskbarVisibility, setTaskbarVisibility, { immediate: true })
